+++ src/ServiceControl.Persistence.RavenDb5/EventLogDataStore.cs
@@ -25,7 +25,7 @@
             }
         }
 
-        public async Task<(IList<EventLogItem>, int, string)> GetEventLogItems(PagingInfo pagingInfo)
+        public async Task<(IList<EventLogItem>, long, string)> GetEventLogItems(PagingInfo pagingInfo)
         {
             using (var session = documentStore.OpenAsyncSession())
             {
